Refactor validation checks
Creates the basis for aborting validation early.
