+++ libs/language-server/src/lib/jayvee-module.ts
@@ -22,36 +22,12 @@ import { registerConstraints } from './constraint/constraint-registry';
 import { JayveeHoverProvider } from './hover/jayvee-hover-provider';
 import { JayveeValueConverter } from './jayvee-value-converter';
 import { JayveeValidationRegistry } from './validation/validation-registry';
-import { BlockValidator } from './validation/validators/block-validator';
-import { CellRangeSelectionValidator } from './validation/validators/cell-range-selection-validator';
-import { ColumnExpressionValidator } from './validation/validators/column-expression-validator';
-import { ConstraintValidator } from './validation/validators/constraint-validator';
-import { ExpressionValidator } from './validation/validators/expression-validator';
-import { JayveeModelValidator } from './validation/validators/model-validator';
-import { PipeValidator } from './validation/validators/pipe-validator';
-import { PipelineDefinitionValidator } from './validation/validators/pipeline-validator';
-import { PropertyBodyValidator } from './validation/validators/property-body-validator';
-import { RegexValueValidator } from './validation/validators/regex-value-validator';
-import { ValuetypeValidator } from './validation/validators/valuetype-validator';
 
 /**
  * Declaration of custom services - add your own service classes here.
  */
-export interface JayveeAddedServices {
-  validation: {
-    JayveeModelValidator: JayveeModelValidator;
-    PipelineDefinitionValidator: PipelineDefinitionValidator;
-    PipeValidator: PipeValidator;
-    BlockValidator: BlockValidator;
-    PropertyBodyValidator: PropertyBodyValidator;
-    CellRangeSelectionValidator: CellRangeSelectionValidator;
-    ColumnExpressionValidator: ColumnExpressionValidator;
-    ConstraintValidator: ConstraintValidator;
-    RegexValueValidator: RegexValueValidator;
-    ValuetypeValidator: ValuetypeValidator;
-    ExpressionValidator: ExpressionValidator;
-  };
-}
+// eslint-disable-next-line @typescript-eslint/no-empty-interface
+export interface JayveeAddedServices {}
 
 /**
  * Union of Langium default services and your custom services - use this as constructor parameter
@@ -73,17 +49,6 @@ export const JayveeModule: Module<
   },
   validation: {
     ValidationRegistry: (services) => new JayveeValidationRegistry(services),
-    JayveeModelValidator: () => new JayveeModelValidator(),
-    PipelineDefinitionValidator: () => new PipelineDefinitionValidator(),
-    PipeValidator: () => new PipeValidator(),
-    BlockValidator: () => new BlockValidator(),
-    PropertyBodyValidator: () => new PropertyBodyValidator(),
-    CellRangeSelectionValidator: () => new CellRangeSelectionValidator(),
-    ColumnExpressionValidator: () => new ColumnExpressionValidator(),
-    ConstraintValidator: () => new ConstraintValidator(),
-    RegexValueValidator: () => new RegexValueValidator(),
-    ValuetypeValidator: () => new ValuetypeValidator(),
-    ExpressionValidator: () => new ExpressionValidator(),
   },
   lsp: {
     CompletionProvider: (services: LangiumServices) =>
