+++ libs/language-server/src/lib/validation/checks/pipe-definition.ts
@@ -1,0 +1,41 @@
+// SPDX-FileCopyrightText: 2023 Friedrich-Alexander-Universitat Erlangen-Nurnberg
+//
+// SPDX-License-Identifier: AGPL-3.0-only
+
+import { ValidationAcceptor } from 'langium';
+
+import { PipeDefinition } from '../../ast/generated/ast';
+import { createSemanticPipes } from '../../ast/wrappers/pipe-wrapper';
+import { getMetaInformation } from '../../meta-information/meta-inf-registry';
+
+export function validatePipeDefinition(
+  pipe: PipeDefinition,
+  accept: ValidationAcceptor,
+): void {
+  checkBlockCompatibility(pipe, accept);
+}
+
+function checkBlockCompatibility(
+  pipe: PipeDefinition,
+  accept: ValidationAcceptor,
+): void {
+  const semanticPipes = createSemanticPipes(pipe);
+  for (const semanticPipe of semanticPipes) {
+    const fromBlockType = semanticPipe.from.type;
+    const toBlockType = semanticPipe.to.type;
+
+    const fromBlockMetaInf = getMetaInformation(fromBlockType);
+    const toBlockMetaInf = getMetaInformation(toBlockType);
+    if (fromBlockMetaInf === undefined || toBlockMetaInf === undefined) {
+      continue;
+    }
+
+    if (fromBlockMetaInf.hasOutput() && toBlockMetaInf.hasInput()) {
+      if (!fromBlockMetaInf.canBeConnectedTo(toBlockMetaInf)) {
+        const errorMessage = `The output type "${fromBlockMetaInf.outputType}" of ${fromBlockMetaInf.type} is incompatible with the input type "${toBlockMetaInf.inputType}" of ${toBlockMetaInf.type}`;
+        accept('error', errorMessage, semanticPipe.getFromDiagnostic());
+        accept('error', errorMessage, semanticPipe.getToDiagnostic());
+      }
+    }
+  }
+}
